// frontend/src/openapi/models/PlaygroundFeatureSchema.ts
export interface PlaygroundFeatureSchema {
    /**
     * The feature's name.
     * @type {string}
     * @memberof PlaygroundFeatureSchema
     */
    name: string;
    /**
     * The ID of the project that contains this feature.
     * @type {string}
     * @memberof PlaygroundFeatureSchema
     */
    projectId: string;
    /**
     * 
     * @type {PlaygroundFeatureSchemaStrategies}
     * @memberof PlaygroundFeatureSchema
     */
    strategies: PlaygroundFeatureSchemaStrategies;
    /**
     * Whether the feature is active and would be evaluated in the provided environment in a normal SDK context.
     * @type {boolean}
     * @memberof PlaygroundFeatureSchema
     */
    isEnabledInCurrentEnvironment: boolean;
    /**
     * Whether this feature is enabled or not in the current environment.
     *                           If a feature can't be fully evaluated (that is, `strategies.result` is `unknown`),
     *                           this will be `false` to align with how client SDKs treat unresolved feature states.
     * @type {boolean}
     * @memberof PlaygroundFeatureSchema
     */
    isEnabled: boolean;
    /**
     * 
     * @type {PlaygroundFeatureSchemaVariant}
     * @memberof PlaygroundFeatureSchema
     */
    variant: PlaygroundFeatureSchemaVariant | null;
    /**
     * 
     * @type {Array<VariantSchema>}
     * @memberof PlaygroundFeatureSchema
     */
    variants: Array<VariantSchema>;
}